requirements.txt
aiocoap==0.4.1
zeroconf==0.29.0
peewee==3.14.4
peewee-async==0.7.1
aiopg==1.2.1
influxdb_client==1.15.0
pika==1.2.0
pyyaml
sqlalchemy
asyncpg